Move dialog to utils
