+++ fract4dgui/utils.py
@@ -106,3 +106,31 @@ class ColorButton(Gtk.ColorButton):
                               color.green / 65535.0, color.blue / 65535.0, 1.0)
         self.changed_cb(self.color.red, self.color.green,
                         self.color.blue, self.is_left)
+
+
+class Dialog(Gtk.Dialog):
+    def __init__(self, title=None, parent=None, buttons=None, modal=Gtk.DialogFlags.MODAL):
+        super().__init__(
+            title=title,
+            transient_for=parent,
+            modal=modal,
+            destroy_with_parent=Gtk.DialogFlags.DESTROY_WITH_PARENT)
+
+        if buttons:
+            self.add_buttons(*buttons)
+
+        self.set_default_response(Gtk.ResponseType.CLOSE)
+        self.connect('response', self.onResponse)
+        self.connect('delete-event', self.quit)
+
+    def onResponse(self, widget, id):
+        if id == Gtk.ResponseType.CLOSE or \
+                id == Gtk.ResponseType.NONE or \
+                id == Gtk.ResponseType.DELETE_EVENT:
+            self.hide()
+        else:
+            print("unexpected response %d" % id)
+
+    def quit(self, widget, event):
+        self.hide()
+        return True
